WebApp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.9
fastapi==0.121.0
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-experimental==0.4.0
langchain-google-genai==3.0.2
langchain-groq==1.0.0
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
numpy==2.3.4
oauthlib==3.3.1
ollama==0.6.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
pip==25.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pymupdf==1.26.6
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
sqlalchemy==2.0.44
sqlite-vec==0.1.6
starlette==0.49.3
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
